tracking_braspress/requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
pyodbc>=5.1.0
